Extract iterable.test.js from common.test.js
